fix(queues): pin task and scheduled-task to BullMQ too
processTask calls processInboundMessage -> noStreamProcess ->
upsertConversationHistory, so a task run on trigger.dev publishes its
assistant-reply row event onto the wrong Redis — the same stuck-on-
"Working…" bug the previous commit fixed for the other three jobs.
scheduled-task is what enqueues task, so the two move together.

initAlwaysOnWorkers is async now: it owns initializeScheduledTaskScheduler,
which recovers scheduled tasks whose nextRunAt passed while the process was
down. That has to run on both providers now that scheduled-task is always
ours, so initWorkers no longer calls it separately. hasScheduledJob loses
its trigger branch for the same reason.

enqueueScheduledTask drops the trigger.dev comment about idempotencyKey
stalls — BullMQ dedupes on jobId, and callers already removeScheduledTask
before re-enqueueing.

Deleted trigger/task/task.ts and the now-unused @trigger.dev/sdk import in
queue-adapter. Every caller goes through the adapter helpers, so no call
sites changed; the task-lifecycle tests mock those helpers and still pass.

Note for trigger.dev deployments: task runs are full agent turns, up to
30 minutes, and they now execute in the webapp process at concurrency 5
(trigger ran them on large-1x at concurrencyLimit 3). Worth watching
webapp memory after this lands.
